Sponsors Page, What, and Who updated
The google map link works. If the map image is clicked, it will bring up
the address to the Compuware building in Google maps.
All of the links on the sponsors page work. But as of now, it doesn't
open the pages in new tabs. I would like to make it open the pages in
new tabs. (Optional)
For the Who page, I am still working on those jump links that Dashaun
specified. The only thing completed on the page is the linking to the
sponsor sites.