Jump Links on "Who" Page, Google Analytics
I re-added Google Analytics, since we didn't have the code in this new
version of the website. It is still in it's local test mode, so before
debut, we will have to change it back to the original code given by
google.
Also, the Jump Links work on the Who page. I added some space saving
boxes like shown in the mockup.
I still can't get a background image to show up for the What page. I
will ask Aziz about this via email.