Still having image issues
I copied your code that you used for the image on the "What you Gain
Page", trying to get something to show up on my "What is IT in the D
page" but it doesn't work when I change it. It will work when I use your
class though. So I'm thinking it has something to do with the way I
copied the class in the css file.
I will demonstrate tomorrow during class. Hopefully we can solve it
quickly together.